+++ src/Symfony/Component/Form/Tests/Command/DebugCommandTest.php
@@ -198,7 +198,7 @@ class FooType extends AbstractType
     {
         $resolver->setRequired('foo');
         $resolver->setDefined('bar');
-        $resolver->setDeprecated('bar');
+        $resolver->setDeprecated('bar', 'vendor/package', '1.1');
         $resolver->setDefault('empty_data', function (Options $options) {
             $foo = $options['foo'];